Fix #4205 (step 6): Alias 'colour' for 'color'
Allow the British English spelling of 'color' (colour) as aliases. Implemented at the command line for `--color`, `--stack-colors` and `stack ls stack-colors`. Implemented in yaml configuration files for `color:` and `stack-colors:` (in each case, the American spelling is the alternative that has priority).

The documentation in `yaml_configuration.md` is updated, accordingly.