Separate out Snapshot and SnapshotLayer
A single snapshot YAML file is in fact just a layer in the creation of a
snapshot. In the current code, Stack will ask Pantry to perform the
parsing of the individual layers, and then combine the layers into its
own representation of a snapshot. This means that the code for this is
less easily testable, and other tooling (e.g., stack2nix) will not have
access to this logic.

This commit separates out the idea of a snapshot and a snapshot layer at
the type level. The next commit will be a simple change to the Stack
codebase to use the updated names, but keep the layer combining logic it
currently has. A future, somewhat more involved change will drop the
duplicated logic from Stack entirely and rely on Pantry's logic
exclusively.