Proper relative paths for script resolvers
Previously, any relative paths in a script's --resolver flag would be
considered relative to the current working directory. It's more logical
and consistent to have it relative to the directory the script lives in.
This makes the change.

Possible argument: it's a bit weird to put the resolver root into the
GlobalOptsMonoid value. Overall it seems like the best option to me, but
I'm open to other approaches.