Modify the GHC_ENVIRONMENT env var in Stack.Setup
This fixes #4706. Since GHC 8.0, GHC will now implicitly read in from a
.ghc.environment.* file, which can cause commands like `stack exec ghci`
to fail. Due to the use case of `stack exec`, it doesn't make sense to
try to create our own environment file, but instead tell GHC to ignore
it. Unfortunately, the ability to ignore environment files was only
added in GHC 8.4.4. This patch:

* Unsets any `GHC_ENVIRONMENT` variable already set outside of Stack
* When using GHC 8.4.4 or later, sets the variable to `-`

This will help work around situations where `cabal new-build` creates an
environment file without the user's awareness. This may be somewhat
superfluous in the future depending on changes to either GHC or
cabal-install, but shouldn't present any harm for the foreseeable future
(unless GHC changes its understanding of `GHC_ENVIRONMENT`).