Set empty input to sinkProcessStderrStdout
This is intended to address intermittent integration test failures,
which started as a result of the new integration test runner. This is a
weird corner case, but here goes:

* ./configure scripts will fail if run with `stdin` closed. We already
work around this in Stack.Build.Execute in the OTLogFile case

* However, in the OTConsole case, we use the sinkProcessStderrStdout
functions, which inherits stdin

* The current behavior can be classified as a bug: we don't want
subprocesses to be exposed to stdin in general.

    * Reviewer request: I've checked the other use cases for
      sinkProcessStderrStdout, and I believe in _all_ cases we do not want the
      subprocess to have such stdin access. Please confirm I haven't made a
      mistake.

* The new integration test runner explicit uses `setStdin closed`...

* Starting with aaa7a547db12e3be8f12b0068f50d0f54fdfed71, we use
interleaved output by default, which causes the OTConsole code path to
occur in integration tests when building the `network` package for the
upload test suite. And since the `network` package has a ./configure
script, its build fails with a file descriptor 0 error