withLockedDir that blocks
Previous implementation errored out when a lock failed. However, it
appears that this is a bit of a race condition within a process: some
integration tests fail because the previous unlock for the same dist dir
hasn't fully synced on the system. This change makes it so that we block
until the lock is available. This not only addresses that problem, but
may have been better behavior in the first place.

While blocking, Stack will print a message every 30 seconds indicating
that the block is still in effect.