Added support for MX linux distribution
MX Linux is based on Debian stable, so it is trivial to add support
for it.

During testing I noticed that the dpkg-query check in
apt_get_install_pkgs always assumes the packages are installed even
when they aren't. This is probably a subtle behavior change between
the /bin/sh versions in MX and Debian testing (I am not entirely sure
though). This check was introduced in commit b46e5a7 a few days
back, reverting that to the previous version works fine in MX and
Debian stretch. Please let me know if this is a problem, I'll try to
find an alternative fix.