Fix #4726: On Windows, warn on 'programs' path with space, no 8dot3 name
In module `Stack.Config`, function `configFromConfigMonoid`, a stack user is warned if the stack 'programs' path contains a space character and does not have an alternative short ('8 dot 3') name,warned of the implications for packages that make use of the GNU project's `configure` shell script.

New modules `System.Info.ShortPathName` are added to `src/windows` and `src/unix` to rexport `getShortPathName` from the `System.Win32.Info` module of the `Win32` package in the case of Windows and do nothing in the case of Unix-like operating systems. This avoids the need for C preprocessor (CPP) directives in `Stack.Config`.

`ChangeLog.md` is updated, accordingly.

Tested on Windows 10 with 'programs' paths that do and do not contain space characters and, in the latter case, with paths that do and do not have '8 dot 3' names. Tested with paths that do not yet exist when stack is first run.