Inhibit errors from terminal-size
When attempting to detect the size of a Windows terminal, stty can
sometimes fail with an error message, which the terminal-size package
then prints to the console. This causes unexpected error messages when
attempting to use Stack in such a terminal. These error messages can
also interfere with tools such as Intero, which rely on Stack's output.