Enable nightly builds again for master branch
@borsboom pinged me that nightly pipeline seems to be missing in the
UI. Analyzing it further, I found that the nightly pipeline in the
dashboard has zero scheduled builds. The likely reason for this seems
to be because of new deployment of azure pipelines.

The good news is now we can control the nightly builds using the yaml
file itself (this wasn't possible before when we added azure builds to
stack initially). This PR does that and make sure we don't rely on UI
for the CI configuration anymore.