Installing specific versions of stack
This is intended to fix issue 5067.

It adds a new section "Install Older Versions" to
the end of doc/install_and_upgrade.md, which tells
users how to find a specific version of stack on
the release page, and also shows how one could add
some lines to their appveyor.yml to download and
install a specific version of stack.