Hacky workaround: run integration tests with rawSystem
We were getting an error when running the integration tests on Windows
due to the presence of a Unicode character in the standard output. From
what I can tell, this is actually a bug in Shake's running of processes.
By changing over to rawSystem, we bypass that codepath and get to see if
that theory is true.