Use released filelock
We've been hoping for https://github.com/takano-akio/filelock/pull/7 to
get merged and released. However, this is dragging on, and we'd rather
use the released version of all packages in releases. Ultimately the
interruptible FFI issue isn't serious enough to warrant continuing to
use the non-released version of filelock.