Do not remove sublibrary information when adding pvp-bounds
This fixes `stack sdist` broken when using pvp-bounds, see
https://github.com/commercialhaskell/stack/issues/5289. This is a
regression after https://github.com/commercialhaskell/stack/pull/5156.

When using pvp-bounds we pretty-print an updated package description
back to .cabal file. If a set of sublibraries used in a dependency is
empty (from cabal-the-library perspective) then `instance Pretty
Dependency` from
https://hackage.haskell.org/package/Cabal-3.0.0.0/docs/src/Distribution.Types.Dependency.html#line-58
will just print an empty set there:

    build-depends: vector : {} <0.13

This leads to the following error when using `stack sdist`:

    - 46:16:
    unexpected Sublibrary dependency syntax used. To use this syntax the package needs to specify at least 'cabal-version: 3.0'. Alternatively, if you are depending on an internal library, you can write directly the library name as it were a package.
    expecting space

What we want instead is to use the same sublibrary set as the original
dep. I suspect most of the time it's just `[LMainLibName]`, which
thanks to `instance Pretty Dependency` omits the `{}` in the
dependency line and is hence (hopefully) compatible with whatever
`cabal-version` the original .cabal file used.