Fix #5322 Extend `stack exec --help` docs and consequent changes
Adds, to `stack exec --help`, `If the command is absent, the first of any arguments is taken as the command.`

Given that change, expands the meta variables to `COMMAND` (from `CMD`) and `ARGUMENT(S)` (from `ARG`). The `(S)` is used to indicate one or more arguments.

Changes the example from `stack exec -- ghc-pkg describe base` to `stack exec ghc-pkg -- describe base`, to follow the alternative actually presented by `stack exec --help`.

In addition, for consistency throughout:

- Changes, throughout, the meta variable of `--package` to `PACKAGE(S)` (from default `ARG`), to avoid confusion over different references to 'arguments'. Also changes `Additional packages` to `Additional package(s)` to indicate one or more packages.

- Makes consistent changes to `stack script --help`.

- Makes consistent changes to `stack build --help` in respect of the `--exec` option.

- Makes consistent changes to `stack hoogle --help`, which referred to `'stack exec' synatax`. Also, rather than cross-refer user to `stack exec`, refers to `the '-- ARGUMENT(S)' syntax` and conforms to the `stack exec --help` approach of giving an example also after `-- ARGUMENT(S)`.