Tidy up and extend CI
This pull request is intended to tidy up the CI and the associated `--stack-yaml` files.

The current `stack.yaml` specifies `lts-17.15` (GHC 8.10.4). As that is after GHC 8.8, I have assumed that `stack-ghc-88.yaml` is not required and the unit tests that refer to it can be safely dropped.

The current `stack-ghc-810.yaml` just refers to `stack.yaml`. I have assumed the corresponding unit tests can refer directly to `stack.yaml` and `stack-ghc-810.yaml` is not required.

The current `stack-nightly.yaml` specifies `nightly-2021-04-30` (GHC 8.10.4). That was first published before `lts-17.15` (also GHC 8.10.4, published on 2021-06-06). I have assumed that the unit tests that refer to it are not required.

`stack` builds with GHC 9.0.2 (lts-19.1). I have added `stack-ghc-902.yaml` to the unit tests for Ubuntu and Windows. See https://github.com/commercialhaskell/stack/pull/5696 and https://github.com/commercialhaskell/stack/issues/5697.

Only minor changes to `stack` are required for it to build also with GHC 9.2.2. I have added a `stack-ghc-922.yaml` in anticipation of that that (which references `nightly-2022-04-09`).