Fix #5713 Don't assume unified report redundant if only one *.tix file
Currently, `Stack.Coverage.generateHpcUnifiedReport` assumes that a unified coverage report will be redundant if there is only one `*.tix` file. However, that is not necessarily the case. For example, one package may test the library of another package that does not test its own library.

As an interim solution, this proposed pull request suggests that a unified report should always be produced if there is one or more `*.tix` files, even if it may be redundant in some circumstances.

A more complex solution would be to have a more complex test that determines whether a unified report would be truely redundant if there is only one `*.tix` file.