Fix #1369 Add build option `--cabal-verbosity=VERBOSITY`
Retains the existing `--[no-]cabal-verbose`, as an alternative flag on the command line.

As with the existing flag, the new option can be set in configuration files. The new option has priority in configuration files.

Error messages are provided if the specified verbosity is not recognised by Cabal (the library).

The `ChangeLog.md` and documentation are updated accordingly.

Unit tests are updated accordingly.

Tested by building Stack and using it with various combinations of options at the command line and/or in a `stack.yaml` file.