Add docs on Stack and VS Code
The use of the 'Haskell' extension for Visual Studio Code (VS Code), the extent of its current integration with Stack, and the possible implications of that for Stack's position relative to other build tools for Haskell (namely, Cabal (the tool)) have been the subject of some extensive discusion recently on the [Haskell Community](https://discourse.haskell.org/), under a number of Topics.

This proposed pull request seeks to add documentation on Stack and VS Code and the current position. Its objective is to capture that there are current viable workarounds to the limitations and that future developments may see the extension able to work well with Stack.