Restructure the Guide between introductory and GUIDE_advanced
Other than typographical corrections, this commit does not change any existing content. All it seeks to do is to separate the existing monolithic Guide web page into the introductory material (consistent with the existing introduction to the Guide) from the more advanced material (characterised in the Guide as 'fun features' or 'power user' commands).

The intent is that this will provide a better foundation for more comprehensive online guidance for the more advanced features.