Increase scope of, and apply, HLint 3.5 suggestions
The hints once considered not useful do appear to be useful when applied to Stack's current code.

With HLint 3.5, some built-in hints are now off-by-default and do not need to be expressly ignored.

For the tests, a different approach is taken to managing the conflict between CPP directives and HLint.

The environment varable `WINDOWS` is no longer used in Stack's code.