Move Stack to -XGHC2021
Merging this now may be premature. Users of Stack < 2.9.1 will not be able to `stack upgrade --source-only --git`, because Hpack >= 0.35.0 is required, although they will be able to `stack upgrade` to a version of Stack that can do that. Stack 2.9.1 was released in September 2022.

On applying GHC2021, `Stack.Build.Execute.withLockedDistDir` required a signature on `complainer` in order to compile.