Fix #6180 Take a different approach to building GHC from source
The `--configure`\\`-c` flag of Hadrian has been deprecated and the feature removed. `Stack.Setup.buildGhcFromSource` no longer uses the flag but seeks to run the GHC repository's `boot` Python script and `configure` sh script.

`Stack.Setup.buildGhcFromSource` requires MSYS2 tools, so - borrowing the pattern of `WithGHC env` - `newtype WithMSYS env` and `runWithMSYS` are defined.

The GHC repository does not provide its `configure` file, so it has to be created by applying `autoreconf`. It also requires `aclocal`. Moved `ensureConfigureScript` out of `Stack.Build.Execute.ensureConfig` to new module `Stack.Config.ConfigureScript` and extended the help on Windows to refer to `aclocal` and the MSYS2 package `autotools`.

GHC's configure script requires `ghc`, `happy` and `alex` on the PATH. So, (1) it is run in Stack's environment, making use of the `stack.yaml` at `hadrian`; and (2) `happy` and `alex` are installed using Stack, if not on the PATH.

It appears that, for some commits, GHC's configure script requires a more recent version of GHC than is specified in the `stack.yaml`. So, if `--resolver` is specified at the command line, that will be used by Stack and can also be specified in the `STACK` environment variable used by Hadrian's `build-stack` scripts.

It also appears that `hadrian` has some other pre-requisites that are not included in the Stack-supplied MSYS2 out of the box and have to be installed if absent. They include `python` and the python package `sphinx`. They also include DejaVu fonts. They are not dealt with here, but the online documentation is updated accordingly.