Bump to Stackage LTS Haskell 23.17
Depends on the existing work around for: https://github.com/commercialhaskell/stack/issues/6711.

On Windows only, for some unidentified reason, stack script can fail when using a pre-compiled package. This affects the script build-stack-installer.hs because, with LTS Haskell 23.17, syb becomes a dependency of Stack and the script sees syb as a pre-compiled package. The work around is to build the package required for that script using the same Stack configuration as used by the script.