Re #6711 Add work around to Stack's CI
Added as part of the work around for: https://github.com/commercialhaskell/stack/issues/6711

On Windows only, for some unidentified reason, stack script can fail when using a pre-compiled package. This can affect the script build-stack-installer.hs. The work around is to build the package required for that script using the same Stack configuration as used by the script.

Adds dummy package description package.yaml. It is required because the first place that Stack looks for a target is project packages and without a Cabal file, Stack will report error [S-636].