Make `stack ghci`'s Stack errors prettier
Also is more accurate about what was `MissingFileTarget` (now `FileTargetIsInvalidAbsFile`). The exception is not always thrown when a specified 'file' is 'missing': it is also thrown when what is specified is not an absolute path to a file and it is not thrown when what is specified is an absolute path.